requirements.txt
PyAutoGUI~=0.9.54
keyboard~=0.13.5
pillow~=10.4.0
art~=6.2
icecream~=2.1.3